requirements.txt
asgiref==3.11.1
Django==6.0.3
django-filter==25.2
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
python-dotenv==1.2.2
sqlparse==0.5.5
